Treesame-Commit-Id: 3657ca42e2d37709fe44a21babba3da64c4526e1

+++ app/models/quizzes/quiz_question.rb
@@ -50,7 +50,7 @@ class Quizzes::QuizQuestion < ActiveRecord::Base
   before_save :create_assessment_question, unless: :generated?
   before_destroy :delete_assessment_question, unless: :generated?
   before_destroy :update_quiz
-  validates :quiz_id, presence: true
+  validates_presence_of :quiz_id
   serialize :question_data
   after_save :update_quiz
 
@@ -73,11 +73,11 @@ class Quizzes::QuizQuestion < ActiveRecord::Base
 
   def infer_defaults
     if !self.position && self.quiz
-      self.position = if self.quiz_group
-                        (self.quiz_group.quiz_questions.active.filter_map(&:position).max || 0) + 1
-                      else
-                        self.quiz.root_entries_max_position + 1
-                      end
+      if self.quiz_group
+        self.position = (self.quiz_group.quiz_questions.active.map(&:position).compact.max || 0) + 1
+      else
+        self.position = self.quiz.root_entries_max_position + 1
+      end
     end
   end
 
@@ -99,10 +99,9 @@ class Quizzes::QuizQuestion < ActiveRecord::Base
   #  The user/teacher who's performing the regrade (e.g, updating the question).
   #  Note that this is NOT an id, but an actual instance of a User model.
   def question_data=(in_data)
-    data = case in_data
-           when String
+    data = if in_data.is_a?(String)
              ActiveSupport::JSON.decode(in_data)
-           when Hash
+           elsif in_data.is_a?(Hash)
              in_data.with_indifferent_access
            else
              in_data
@@ -143,25 +142,25 @@ class Quizzes::QuizQuestion < ActiveRecord::Base
   end
 
   def delete_assessment_question
-    if self.assessment_question&.editable_by?(self)
+    if self.assessment_question && self.assessment_question.editable_by?(self)
       self.assessment_question.destroy
     end
   end
 
   def create_assessment_question
-    return if self.question_data&.is_type?(:text_only)
+    return if self.question_data && self.question_data.is_type?(:text_only)
 
     aq = self.assessment_question || AssessmentQuestion.new
 
     if aq.editable_by?(self)
       aq.question_data = self.question_data
-      aq.initial_context = self.quiz.context if self.quiz&.context
+      aq.initial_context = self.quiz.context if self.quiz && self.quiz.context
       aq.save! if aq.new_record?
     end
 
     self.assessment_question = aq
 
-    true
+    return true
   end
 
   def update_assessment_question!(aq, quiz_group_id, duplicate_index)
@@ -173,14 +172,14 @@ class Quizzes::QuizQuestion < ActiveRecord::Base
     self.duplicate_index = duplicate_index
     save! if changed?
 
-    self
+    return self
   end
 
   def validate_blank_questions
     return if self.question_data && !(self.question_data.is_type?(:fill_in_multiple_blanks) || self.question_data.is_type?(:short_answer))
 
     qd = self.question_data
-    qd.answers = qd.answers.reject { |answer| answer['text'].empty? }
+    qd.answers = qd.answers.select { |answer| !answer['text'].empty? }
     self.question_data = qd
     self.question_data_will_change!
     true
@@ -218,7 +217,7 @@ class Quizzes::QuizQuestion < ActiveRecord::Base
   # All questions will be assigned to the given quiz_group, and will be
   # assigned as part of the root quiz if no group is given
   def self.update_all_positions!(questions, quiz_group = nil)
-    return if questions.empty?
+    return unless questions.size > 0
 
     group_id = quiz_group ? quiz_group.id : 'NULL'
     updates = questions.map do |q|
